% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aTPR.R
\name{aTPR}
\alias{aTPR}
\title{Calibrate risk}
\usage{
aTPR(
  data,
  groupvar,
  ref,
  response,
  risk,
  taus,
  calmethod = c("logit"),
  cal.args = list(2),
  drmethod = c("logit"),
  dr.args = list(2),
  cv = FALSE
)
}
\arguments{
\item{data}{a data.frame containing the original and re-calibrate risks, density ratio estimate and group label}

\item{groupvar}{the group column}

\item{ref}{reference level}

\item{response}{response variable}

\item{risk}{the data.frame column representing the (original) risk score under evaluation}

\item{taus}{a vector containing the decision thresholds of interest}

\item{calmethod}{method to be used for calibration, default "logit"}

\item{cal.args}{additional arguments for calibration method; with "logit" degree of polynomial}

\item{drmethod}{method for estimating density ratio}

\item{dr.args}{arguments for density ratio estimation method}

\item{cv}{logical indicating whether MSPE should be estimated by cross-validation}
}
\value{
a list with taus and aTPRs
}
\description{
Calculate the re-calibrated risk within each group.
}
