% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{calibrate}
\alias{calibrate}
\title{calibrate}
\usage{
calibrate(train, test, method, args)
}
\arguments{
\item{train}{a data.frame containing the risk score, observed response and group label}

\item{test}{label, categorical}

\item{method}{method to be used to calibrate the risks, options: 'logit' for logistic regression}

\item{args}{optional odditional arguments to be used for the method. e.g., when method = 'logit', the degree of the polynomial can be specified}
}
\value{
data.frame with the calibrated risk results
}
\description{
Inner function for re-calibrated risk within a group.
}
