% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrateRiskCV.R
\name{calibrateRiskCV}
\alias{calibrateRiskCV}
\title{Calibrate risk}
\usage{
calibrateRiskCV(
  data,
  groupvar,
  response,
  risk,
  transform = TRUE,
  method = "qlogit",
  args = list(2),
  cv = TRUE,
  k = 5
)
}
\arguments{
\item{data}{a data.frame containing the risk score, observed response and group label}

\item{groupvar}{the group label, categorical}

\item{response}{the observed outcome or response variable, should be binary 0/1}

\item{risk}{the (original) risk score under evaluation, scalar}

\item{transform}{apply a logit transformation to the risk score, converts probability to a linear predictor}

\item{method}{method to be used to calibrate the risks, options: 'logit' for logistic regression}

\item{args}{optional odditional arguments to be used for the method. e.g., when method = 'logit', the degree of the polynomial can be specified}

\item{cv}{logical indicating if cross-validation estimate of MSPE should be performed}

\item{k}{number of folds for cross-validation, default = 5}
}
\value{
data.frame with the calibrated risk results
}
\description{
Calculate the re-calibrated risk within each group.
}
