% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estDensityRatio.R
\name{estDensityRatio}
\alias{estDensityRatio}
\title{Estimate density ratio}
\usage{
estDensityRatio(
  train,
  test,
  method = "logit",
  args = list(1),
  groupvar = .data$s,
  refgp = "1",
  calrisk = .data$rs.gX,
  quietly = TRUE
)
}
\arguments{
\item{train}{a data.frame containing the risk score, observed response and group label}

\item{test}{data.frame containing the re-calibrated risk score, and group label}

\item{method}{method to be used to calibrate the risks, options: 'logit' for logistic regression}

\item{args}{optional odditional arguments to be used for the method. e.g., when method = 'logit', the degree of the polynomial can be specified}

\item{groupvar}{group variable}

\item{refgp}{reference group level}

\item{calrisk}{re-calibrated risk score}

\item{quietly}{logical for suppression of output messages, default = TRUE}
}
\value{
a data.frame with the estimated density ratio weights
}
\description{
Calculate the density ratio of the re-calibrated risk between each group and reference group.
}
