% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estDensityRatioCV.R
\name{estDensityRatioCV}
\alias{estDensityRatioCV}
\title{Estimate density ratio}
\usage{
estDensityRatioCV(
  data,
  method = "logit",
  args = list(1),
  groupvar = .data$s,
  refgp = "1",
  calrisk = .data$rs.gX,
  k = 5
)
}
\arguments{
\item{data}{a data.frame containing the risk score, observed response and group label}

\item{method}{method to be used to calibrate the risks, options: 'logit' for logistic regression}

\item{args}{optional odditional arguments to be used for the method. e.g., when method = 'logit', the degree of the polynomial can be specified}

\item{groupvar}{group variable}

\item{refgp}{reference group level}

\item{calrisk}{calibrated risk score}

\item{k}{number of folds for cross-validation}
}
\value{
a list with taus and aTPRs
}
\description{
Calculate the re-calibrated risk within each group.
}
