% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fairRisk-package.R
\docType{package}
\name{fairRisk-package}
\alias{fairRisk}
\alias{fairRisk-package}
\title{fairRisk: Risk adjusted error rates for evaluating algorithm fairness}
\description{
This package implements the adjusted true positive rate estimator proposed by Hegarty et al.
}
\author{
\strong{Maintainer}: Sarah Hegarty \email{hegarty@upenn.edu} (\href{https://orcid.org/0000-0001-5348-350X}{ORCID})

}
\keyword{internal}
