% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aTNR.R
\name{get_aTNR}
\alias{get_aTNR}
\title{Adjusted True Negative Rate (aTNR)}
\usage{
get_aTNR(
  data,
  orig_risk,
  cal_risk,
  dens_ratio,
  groupvar,
  taus = seq(0.1, 0.9, 0.1)
)
}
\arguments{
\item{data}{a data.frame containing the original and re-calibrate risks, density ratio estimate and group label}

\item{orig_risk}{the data.frame column representing the (original) risk score under evaluation}

\item{cal_risk}{the data.frame column representing the re-calibrated risk score}

\item{dens_ratio}{the data.frame column representing the density ratio of the recalibrated risk scores}

\item{groupvar}{the group column}

\item{taus}{a vector containing the decision thresholds of interest}
}
\value{
a list with taus and aTNRs
}
\description{
Calculate the adjusted true negative rate with respect to a specified reference group.
}
