% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_naivePPV.R
\name{get_naivePPV}
\alias{get_naivePPV}
\title{Standard Positive Predictive Value (PPV)}
\usage{
get_naivePPV(data, risk, response, groupvar, taus = seq(0.1, 0.9, 0.1))
}
\arguments{
\item{data}{a data.frame containing the original and re-calibrate risks, density ratio estimate and group label}

\item{risk}{the data.frame column representing the (original) risk score under evaluation}

\item{response}{the response variable}

\item{groupvar}{the group column}

\item{taus}{a vector containing the decision thresholds of interest}
}
\value{
a list with taus and PPVs
}
\description{
Calculate the positive predictive value
}
