% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naiveTNR.R
\name{naiveTNR}
\alias{naiveTNR}
\title{Standard True Negative Rate (TNR)}
\usage{
naiveTNR(
  data,
  risk,
  response,
  groupvar,
  taus = seq(0.1, 0.9, 0.1),
  se.boot = FALSE,
  bootsize = 500,
  alpha = 0.05
)
}
\arguments{
\item{data}{a data.frame containing a risk score, response label and group label}

\item{risk}{the data.frame column representing the (original) risk score under evaluation}

\item{response}{the response variable}

\item{groupvar}{the group column}

\item{taus}{a vector containing the decision thresholds of interest}

\item{se.boot}{logical indicating whether bootstrapping should be used to estimate standard error}

\item{bootsize}{number of bootstrapped data sets to sample, default = 500}

\item{alpha}{confidence level for bootstrap quantiles}
}
\value{
a list with a data.frame of TNRs by tau and group and data.frame of bootstrapped estimates if se.boot = TRUE
}
\description{
Calculate the true negative rate
}
