% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampdata.R
\docType{data}
\name{sampdata}
\alias{sampdata}
\title{sample data}
\format{
\code{"sampdata"} is a data frame with 2000 cases (rows) and 5 variables
(columns) named s, pi_s, y, logit.gX, gX
}
\usage{
data(sampdata)
}
\description{
data set containing two groups
}
\keyword{datasets}
