% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aTPR.R
\name{aTPR}
\alias{aTPR}
\title{aTPR}
\usage{
aTPR(
  data,
  groupvar,
  ref,
  response,
  risk,
  taus,
  calmethod = c("logit"),
  cal.args = list(2, TRUE),
  drmethod = c("logit"),
  dr.args = list(2),
  cv = FALSE,
  se.boot = FALSE,
  bootsize = 500,
  alpha = 0.05,
  quietly = TRUE
)
}
\arguments{
\item{data}{a data.frame containing the original and re-calibrated risks, density ratio estimate and group label}

\item{groupvar}{the group column}

\item{ref}{reference level}

\item{response}{response variable}

\item{risk}{the data.frame column representing the (original) risk score under evaluation}

\item{taus}{a vector containing the decision thresholds of interest}

\item{calmethod}{method to be used for calibration, default "logit"}

\item{cal.args}{additional arguments for calibration method; with "logit" degree of polynomial and if risk score should be transformed with logit}

\item{drmethod}{method for estimating density ratio}

\item{dr.args}{arguments for density ratio estimation method}

\item{cv}{logical indicating whether MSPE should be estimated by cross-validation}

\item{se.boot}{logical indicating whether bootstrapping should be used to estimate standard error}

\item{bootsize}{number of bootstrapped data sets to sample, default = 500}

\item{alpha}{confidence level for bootstrap quantiles}

\item{quietly}{suppress messages, default = TRUE}
}
\value{
a list containing a data.frama of aTPR estimates for each tau and group and separate data.frame of bootstrapped estimates if se.boot = TRUE
}
\description{
Calculate the adjusted true positive rate with respect to a specified reference group.
}
