% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estDensityRatioCV.R
\name{estDensityRatioCV}
\alias{estDensityRatioCV}
\title{Estimate density ratio}
\usage{
estDensityRatioCV(
  data,
  method = "logit",
  args = list(1),
  groupvar = .data$s,
  refgp = "1",
  calrisk = .data$rs.gX,
  transform = FALSE,
  cv = TRUE,
  k = 5,
  quietly = TRUE
)
}
\arguments{
\item{data}{a data.frame containing the risk score, observed response and group label}

\item{method}{method to be used to calibrate the risks, options: 'logit' for logistic regression}

\item{args}{optional odditional arguments to be used for the method. e.g., when method = 'logit', the degree of the polynomial can be specified}

\item{groupvar}{group variable}

\item{refgp}{reference group level}

\item{calrisk}{calibrated risk score}

\item{transform}{use logit of calibrated risk (TRUE) or calibrated risk on probability scale (FALSE), default}

\item{cv}{logical indicating if cross-validation should be performed to estimate MSPE}

\item{k}{number of folds for cross-validation}

\item{quietly}{suppress messages, default = TRUE}
}
\value{
a list with data.frame with estimated density ratio weights and MSPE if cv options selected
}
\description{
Calculate the density ratio of the re-calibrated risk between each group and reference group.
}
