% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fairRisk-package.R
\docType{package}
\name{fairRisk-package}
\alias{fairRisk}
\alias{fairRisk-package}
\title{fairRisk: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Sarah Hegarty \email{hegarty@upenn.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
